requirements.lock.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.13.3
betterproto==2.0.0b6
bidict==0.23.1
blinker==1.9.0
bptf==0.1.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==42.0.8
dnspython==2.7.0
Flask==3.1.0
frozenlist==1.5.0
grpclib==0.4.7
h11==0.14.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.3.2
packaging==24.2
pluggy==1.5.0
propcache==0.3.1
pycparser==2.22
pymongo==4.11.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
soupsieve==2.6
steamio==1.1.1
tf2-data==1.0.0
tf2-sku==2.0.1
tf2-utils==2.3.2
typing_extensions==4.13.1
urllib3==2.3.0
vdf==3.4
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
